Round costs to nice numbers

+++ src/main/kotlin/com/davidm1a2/afraidofthedark/common/spell/component/powerSource/ExperienceSpellPowerSource.kt
@@ -27,8 +27,8 @@ class ExperienceSpellPowerSource : AOTDSpellPowerSource(ResourceLocation(Constan
         }
     }
 
-    override fun getSourceSpecificCost(rawCost: Double): Double {
-        return ceil(rawCost / UNIT_COST_PER_XP)
+    override fun getSourceSpecificCost(rawCost: Double): Int {
+        return ceil(rawCost / UNIT_COST_PER_XP).toInt()
     }
 
     companion object {
